add configurer
